+++ pkg/controller/operators/catalog/subscription/reconciler_test.go
@@ -20,7 +20,6 @@ import (
 	registryreconciler "github.com/operator-framework/operator-lifecycle-manager/pkg/controller/registry/reconciler"
 	olmfakes "github.com/operator-framework/operator-lifecycle-manager/pkg/fakes"
 	"github.com/operator-framework/operator-lifecycle-manager/pkg/lib/clientfake"
-	"github.com/operator-framework/operator-lifecycle-manager/pkg/lib/kubestate"
 	"github.com/operator-framework/operator-lifecycle-manager/pkg/lib/operatorlister"
 )
 
@@ -34,11 +33,11 @@ func TestCatalogHealthReconcile(t *testing.T) {
 		config *fakeReconcilerConfig
 	}
 	type args struct {
-		in kubestate.State
+		in *v1alpha1.Subscription
 	}
 	type want struct {
 		err error
-		out kubestate.State
+		out *v1alpha1.Subscription
 	}
 
 	tests := []struct {
@@ -67,39 +66,35 @@ func TestCatalogHealthReconcile(t *testing.T) {
 				},
 			},
 			args: args{
-				in: newSubscriptionExistsState(
-					&v1alpha1.Subscription{
-						ObjectMeta: metav1.ObjectMeta{
-							Name:      "sub",
-							Namespace: "default",
-						},
-						Spec:   &v1alpha1.SubscriptionSpec{},
-						Status: v1alpha1.SubscriptionStatus{},
+				in: &v1alpha1.Subscription{
+					ObjectMeta: metav1.ObjectMeta{
+						Name:      "sub",
+						Namespace: "default",
 					},
-				),
+					Spec:   &v1alpha1.SubscriptionSpec{},
+					Status: v1alpha1.SubscriptionStatus{},
+				},
 			},
 			want: want{
-				out: newCatalogUnhealthyState(
-					&v1alpha1.Subscription{
-						ObjectMeta: metav1.ObjectMeta{
-							Name:      "sub",
-							Namespace: "default",
-						},
-						Spec: &v1alpha1.SubscriptionSpec{},
-						Status: v1alpha1.SubscriptionStatus{
-							Conditions: []v1alpha1.SubscriptionCondition{
-								{
-									Type:               v1alpha1.SubscriptionCatalogSourcesUnhealthy,
-									Status:             corev1.ConditionTrue,
-									Reason:             v1alpha1.NoCatalogSourcesFound,
-									Message:            "dependency resolution requires at least one catalogsource",
-									LastTransitionTime: &now,
-								},
-							},
-							LastUpdated: now,
-						},
+				out: &v1alpha1.Subscription{
+					ObjectMeta: metav1.ObjectMeta{
+						Name:      "sub",
+						Namespace: "default",
 					},
-				),
+					Spec: &v1alpha1.SubscriptionSpec{},
+					Status: v1alpha1.SubscriptionStatus{
+						Conditions: []v1alpha1.SubscriptionCondition{
+							{
+								Type:               v1alpha1.SubscriptionCatalogSourcesUnhealthy,
+								Status:             corev1.ConditionTrue,
+								Reason:             v1alpha1.NoCatalogSourcesFound,
+								Message:            "dependency resolution requires at least one catalogsource",
+								LastTransitionTime: &now,
+							},
+						},
+						LastUpdated: now,
+					},
+				},
 			},
 		},
 		{
@@ -137,21 +132,19 @@ func TestCatalogHealthReconcile(t *testing.T) {
 				},
 			},
 			args: args{
-				in: newSubscriptionExistsState(
-					&v1alpha1.Subscription{
-						ObjectMeta: metav1.ObjectMeta{
-							Name:      "sub",
-							Namespace: "ns",
-						},
-						Spec: &v1alpha1.SubscriptionSpec{
-							CatalogSourceNamespace: "ns",
-							CatalogSource:          "cs-0",
-						},
+				in: &v1alpha1.Subscription{
+					ObjectMeta: metav1.ObjectMeta{
+						Name:      "sub",
+						Namespace: "ns",
 					},
-				),
+					Spec: &v1alpha1.SubscriptionSpec{
+						CatalogSourceNamespace: "ns",
+						CatalogSource:          "cs-0",
+					},
+				},
 			},
 			want: want{
-				out: newCatalogUnhealthyState(&v1alpha1.Subscription{
+				out: &v1alpha1.Subscription{
 					ObjectMeta: metav1.ObjectMeta{
 						Name:      "sub",
 						Namespace: "ns",
@@ -169,7 +162,7 @@ func TestCatalogHealthReconcile(t *testing.T) {
 						},
 						LastUpdated: now,
 					},
-				}),
+				},
 			},
 		},
 		{
@@ -208,7 +201,7 @@ func TestCatalogHealthReconcile(t *testing.T) {
 				},
 			},
 			args: args{
-				in: newSubscriptionExistsState(&v1alpha1.Subscription{
+				in: &v1alpha1.Subscription{
 					ObjectMeta: metav1.ObjectMeta{
 						Name:      "sub",
 						Namespace: "ns",
@@ -227,10 +220,10 @@ func TestCatalogHealthReconcile(t *testing.T) {
 						},
 						LastUpdated: earlier,
 					},
-				}),
+				},
 			},
 			want: want{
-				out: newCatalogUnhealthyState(&v1alpha1.Subscription{
+				out: &v1alpha1.Subscription{
 					ObjectMeta: metav1.ObjectMeta{
 						Name:      "sub",
 						Namespace: "ns",
@@ -249,7 +242,7 @@ func TestCatalogHealthReconcile(t *testing.T) {
 						},
 						LastUpdated: earlier,
 					},
-				}),
+				},
 			},
 		},
 		{
@@ -288,7 +281,7 @@ func TestCatalogHealthReconcile(t *testing.T) {
 				},
 			},
 			args: args{
-				in: newSubscriptionExistsState(&v1alpha1.Subscription{
+				in: &v1alpha1.Subscription{
 					ObjectMeta: metav1.ObjectMeta{
 						Name:      "sub",
 						Namespace: "ns",
@@ -307,10 +300,10 @@ func TestCatalogHealthReconcile(t *testing.T) {
 						},
 						LastUpdated: earlier,
 					},
-				}),
+				},
 			},
 			want: want{
-				out: newCatalogHealthyState(&v1alpha1.Subscription{
+				out: &v1alpha1.Subscription{
 					ObjectMeta: metav1.ObjectMeta{
 						Name:      "sub",
 						Namespace: "ns",
@@ -329,7 +322,7 @@ func TestCatalogHealthReconcile(t *testing.T) {
 						},
 						LastUpdated: earlier,
 					},
-				}),
+				},
 			},
 		},
 		{
@@ -369,7 +362,7 @@ func TestCatalogHealthReconcile(t *testing.T) {
 				},
 			},
 			args: args{
-				in: newSubscriptionExistsState(&v1alpha1.Subscription{
+				in: &v1alpha1.Subscription{
 					ObjectMeta: metav1.ObjectMeta{
 						Name:      "sub",
 						Namespace: "ns",
@@ -388,10 +381,10 @@ func TestCatalogHealthReconcile(t *testing.T) {
 						},
 						LastUpdated: earlier,
 					},
-				}),
+				},
 			},
 			want: want{
-				out: newCatalogHealthyState(&v1alpha1.Subscription{
+				out: &v1alpha1.Subscription{
 					ObjectMeta: metav1.ObjectMeta{
 						Name:      "sub",
 						Namespace: "ns",
@@ -411,7 +404,7 @@ func TestCatalogHealthReconcile(t *testing.T) {
 						},
 						LastUpdated: now,
 					},
-				}),
+				},
 			},
 		},
 		{
@@ -458,30 +451,28 @@ func TestCatalogHealthReconcile(t *testing.T) {
 				},
 			},
 			args: args{
-				in: newSubscriptionExistsState(
-					&v1alpha1.Subscription{
-						ObjectMeta: metav1.ObjectMeta{
-							Name:      "sub",
-							Namespace: "ns",
-						},
-						Spec: &v1alpha1.SubscriptionSpec{
-							CatalogSourceNamespace: "ns",
-							CatalogSource:          "cs-0",
-						},
-						Status: v1alpha1.SubscriptionStatus{
-							CatalogHealth: []v1alpha1.SubscriptionCatalogHealth{
-								catalogHealth("ns", "cs-0", &earlier, true),
-							},
-							Conditions: []v1alpha1.SubscriptionCondition{
-								catalogUnhealthyCondition(corev1.ConditionFalse, v1alpha1.AllCatalogSourcesHealthy, "all available catalogsources are healthy", &earlier),
-							},
-							LastUpdated: earlier,
-						},
+				in: &v1alpha1.Subscription{
+					ObjectMeta: metav1.ObjectMeta{
+						Name:      "sub",
+						Namespace: "ns",
 					},
-				),
+					Spec: &v1alpha1.SubscriptionSpec{
+						CatalogSourceNamespace: "ns",
+						CatalogSource:          "cs-0",
+					},
+					Status: v1alpha1.SubscriptionStatus{
+						CatalogHealth: []v1alpha1.SubscriptionCatalogHealth{
+							catalogHealth("ns", "cs-0", &earlier, true),
+						},
+						Conditions: []v1alpha1.SubscriptionCondition{
+							catalogUnhealthyCondition(corev1.ConditionFalse, v1alpha1.AllCatalogSourcesHealthy, "all available catalogsources are healthy", &earlier),
+						},
+						LastUpdated: earlier,
+					},
+				},
 			},
 			want: want{
-				out: newCatalogUnhealthyState(&v1alpha1.Subscription{
+				out: &v1alpha1.Subscription{
 					ObjectMeta: metav1.ObjectMeta{
 						Name:      "sub",
 						Namespace: "ns",
@@ -499,7 +490,7 @@ func TestCatalogHealthReconcile(t *testing.T) {
 						},
 						LastUpdated: now,
 					},
-				}),
+				},
 			},
 		},
 	}
@@ -513,12 +504,6 @@ func TestCatalogHealthReconcile(t *testing.T) {
 			out, err := rec.Reconcile(ctx, tt.args.in)
 			require.Equal(t, tt.want.err, err)
 			require.Equal(t, tt.want.out, out)
-
-			// Ensure the client's view of the subscription matches the typestate's
-			sub := out.(SubscriptionState).Subscription()
-			clusterSub, err := rec.client.OperatorsV1alpha1().Subscriptions(sub.GetNamespace()).Get(context.TODO(), sub.GetName(), metav1.GetOptions{})
-			require.NoError(t, err)
-			require.Equal(t, sub, clusterSub)
 		})
 	}
 }
@@ -533,11 +518,11 @@ func TestInstallPlanReconcile(t *testing.T) {
 		config *fakeReconcilerConfig
 	}
 	type args struct {
-		in kubestate.State
+		in *v1alpha1.Subscription
 	}
 	type want struct {
 		err error
-		out kubestate.State
+		out *v1alpha1.Subscription
 	}
 
 	tests := []struct {
@@ -564,20 +549,20 @@ func TestInstallPlanReconcile(t *testing.T) {
 				},
 			},
 			args: args{
-				in: newSubscriptionExistsState(&v1alpha1.Subscription{
+				in: &v1alpha1.Subscription{
 					ObjectMeta: metav1.ObjectMeta{
 						Name:      "sub",
 						Namespace: "ns",
 					},
-				}),
+				},
 			},
 			want: want{
-				out: newNoInstallPlanReferencedState(&v1alpha1.Subscription{
+				out: &v1alpha1.Subscription{
 					ObjectMeta: metav1.ObjectMeta{
 						Name:      "sub",
 						Namespace: "ns",
 					},
-				}),
+				},
 			},
 		},
 		{
@@ -608,7 +593,7 @@ func TestInstallPlanReconcile(t *testing.T) {
 				},
 			},
 			args: args{
-				in: newCatalogHealthyState(&v1alpha1.Subscription{
+				in: &v1alpha1.Subscription{
 					ObjectMeta: metav1.ObjectMeta{
 						Name:      "sub",
 						Namespace: "ns",
@@ -623,27 +608,23 @@ func TestInstallPlanReconcile(t *testing.T) {
 						},
 						LastUpdated: earlier,
 					},
-				}),
+				},
 			},
 			want: want{
-				out: &noInstallPlanReferencedState{
-					InstallPlanState: &installPlanState{
-						SubscriptionExistsState: newCatalogHealthyState(&v1alpha1.Subscription{
-							ObjectMeta: metav1.ObjectMeta{
-								Name:      "sub",
-								Namespace: "ns",
-							},
-							Status: v1alpha1.SubscriptionStatus{
-								CatalogHealth: []v1alpha1.SubscriptionCatalogHealth{
-									catalogHealth("ns", "cs-0", &earlier, true),
-									catalogHealth("ns", "cs-1", &earlier, true),
-								},
-								Conditions: []v1alpha1.SubscriptionCondition{
-									catalogUnhealthyCondition(corev1.ConditionFalse, v1alpha1.AllCatalogSourcesHealthy, "all available catalogsources are healthy", &earlier),
-								},
-								LastUpdated: earlier,
-							},
-						}),
+				out: &v1alpha1.Subscription{
+					ObjectMeta: metav1.ObjectMeta{
+						Name:      "sub",
+						Namespace: "ns",
+					},
+					Status: v1alpha1.SubscriptionStatus{
+						CatalogHealth: []v1alpha1.SubscriptionCatalogHealth{
+							catalogHealth("ns", "cs-0", &earlier, true),
+							catalogHealth("ns", "cs-1", &earlier, true),
+						},
+						Conditions: []v1alpha1.SubscriptionCondition{
+							catalogUnhealthyCondition(corev1.ConditionFalse, v1alpha1.AllCatalogSourcesHealthy, "all available catalogsources are healthy", &earlier),
+						},
+						LastUpdated: earlier,
 					},
 				},
 			},
@@ -673,7 +654,7 @@ func TestInstallPlanReconcile(t *testing.T) {
 				},
 			},
 			args: args{
-				in: newSubscriptionExistsState(&v1alpha1.Subscription{
+				in: &v1alpha1.Subscription{
 					ObjectMeta: metav1.ObjectMeta{
 						Name:      "sub",
 						Namespace: "ns",
@@ -685,10 +666,10 @@ func TestInstallPlanReconcile(t *testing.T) {
 						},
 						LastUpdated: earlier,
 					},
-				}),
+				},
 			},
 			want: want{
-				out: newInstallPlanMissingState(&v1alpha1.Subscription{
+				out: &v1alpha1.Subscription{
 					ObjectMeta: metav1.ObjectMeta{
 						Name:      "sub",
 						Namespace: "ns",
@@ -703,7 +684,7 @@ func TestInstallPlanReconcile(t *testing.T) {
 						},
 						LastUpdated: now,
 					},
-				}),
+				},
 			},
 		},
 		{
@@ -734,7 +715,7 @@ func TestInstallPlanReconcile(t *testing.T) {
 				},
 			},
 			args: args{
-				in: newSubscriptionExistsState(&v1alpha1.Subscription{
+				in: &v1alpha1.Subscription{
 					ObjectMeta: metav1.ObjectMeta{
 						Name:      "sub",
 						Namespace: "ns",
@@ -749,10 +730,10 @@ func TestInstallPlanReconcile(t *testing.T) {
 						},
 						LastUpdated: earlier,
 					},
-				}),
+				},
 			},
 			want: want{
-				out: newInstallPlanMissingState(&v1alpha1.Subscription{
+				out: &v1alpha1.Subscription{
 					ObjectMeta: metav1.ObjectMeta{
 						Name:      "sub",
 						Namespace: "ns",
@@ -767,7 +748,7 @@ func TestInstallPlanReconcile(t *testing.T) {
 						},
 						LastUpdated: earlier,
 					},
-				}),
+				},
 			},
 		},
 		{
@@ -798,7 +779,7 @@ func TestInstallPlanReconcile(t *testing.T) {
 				},
 			},
 			args: args{
-				in: newCatalogHealthyState(&v1alpha1.Subscription{
+				in: &v1alpha1.Subscription{
 					ObjectMeta: metav1.ObjectMeta{
 						Name:      "sub",
 						Namespace: "ns",
@@ -817,36 +798,28 @@ func TestInstallPlanReconcile(t *testing.T) {
 						},
 						LastUpdated: earlier,
 					},
-				}),
+				},
 			},
 			want: want{
-				out: &installPlanMissingState{
-					InstallPlanKnownState: &installPlanKnownState{
-						InstallPlanReferencedState: &installPlanReferencedState{
-							InstallPlanState: &installPlanState{
-								SubscriptionExistsState: newCatalogHealthyState(&v1alpha1.Subscription{
-									ObjectMeta: metav1.ObjectMeta{
-										Name:      "sub",
-										Namespace: "ns",
-									},
-									Status: v1alpha1.SubscriptionStatus{
-										InstallPlanRef: &corev1.ObjectReference{
-											Namespace: "ns",
-											Name:      "ip",
-										},
-										CatalogHealth: []v1alpha1.SubscriptionCatalogHealth{
-											catalogHealth("ns", "cs-0", &earlier, true),
-											catalogHealth("ns", "cs-1", &earlier, true),
-										},
-										Conditions: []v1alpha1.SubscriptionCondition{
-											catalogUnhealthyCondition(corev1.ConditionFalse, v1alpha1.AllCatalogSourcesHealthy, "all available catalogsources are healthy", &earlier),
-											planMissingCondition(corev1.ConditionTrue, v1alpha1.ReferencedInstallPlanNotFound, "", &now),
-										},
-										LastUpdated: now,
-									},
-								}),
-							},
+				out: &v1alpha1.Subscription{
+					ObjectMeta: metav1.ObjectMeta{
+						Name:      "sub",
+						Namespace: "ns",
+					},
+					Status: v1alpha1.SubscriptionStatus{
+						InstallPlanRef: &corev1.ObjectReference{
+							Namespace: "ns",
+							Name:      "ip",
 						},
+						CatalogHealth: []v1alpha1.SubscriptionCatalogHealth{
+							catalogHealth("ns", "cs-0", &earlier, true),
+							catalogHealth("ns", "cs-1", &earlier, true),
+						},
+						Conditions: []v1alpha1.SubscriptionCondition{
+							catalogUnhealthyCondition(corev1.ConditionFalse, v1alpha1.AllCatalogSourcesHealthy, "all available catalogsources are healthy", &earlier),
+							planMissingCondition(corev1.ConditionTrue, v1alpha1.ReferencedInstallPlanNotFound, "", &now),
+						},
+						LastUpdated: now,
 					},
 				},
 			},
@@ -877,7 +850,7 @@ func TestInstallPlanReconcile(t *testing.T) {
 				},
 			},
 			args: args{
-				in: newSubscriptionExistsState(&v1alpha1.Subscription{
+				in: &v1alpha1.Subscription{
 					ObjectMeta: metav1.ObjectMeta{
 						Name:      "sub",
 						Namespace: "ns",
@@ -889,10 +862,10 @@ func TestInstallPlanReconcile(t *testing.T) {
 						},
 						LastUpdated: earlier,
 					},
-				}),
+				},
 			},
 			want: want{
-				out: newInstallPlanPendingState(&v1alpha1.Subscription{
+				out: &v1alpha1.Subscription{
 					ObjectMeta: metav1.ObjectMeta{
 						Name:      "sub",
 						Namespace: "ns",
@@ -907,7 +880,7 @@ func TestInstallPlanReconcile(t *testing.T) {
 						},
 						LastUpdated: now,
 					},
-				}),
+				},
 			},
 		},
 		{
@@ -941,7 +914,7 @@ func TestInstallPlanReconcile(t *testing.T) {
 				},
 			},
 			args: args{
-				in: newSubscriptionExistsState(&v1alpha1.Subscription{
+				in: &v1alpha1.Subscription{
 					ObjectMeta: metav1.ObjectMeta{
 						Name:      "sub",
 						Namespace: "ns",
@@ -953,10 +926,10 @@ func TestInstallPlanReconcile(t *testing.T) {
 						},
 						LastUpdated: earlier,
 					},
-				}),
+				},
 			},
 			want: want{
-				out: newInstallPlanPendingState(&v1alpha1.Subscription{
+				out: &v1alpha1.Subscription{
 					ObjectMeta: metav1.ObjectMeta{
 						Name:      "sub",
 						Namespace: "ns",
@@ -971,7 +944,7 @@ func TestInstallPlanReconcile(t *testing.T) {
 						},
 						LastUpdated: now,
 					},
-				}),
+				},
 			},
 		},
 		{
@@ -1008,7 +981,7 @@ func TestInstallPlanReconcile(t *testing.T) {
 				},
 			},
 			args: args{
-				in: newSubscriptionExistsState(&v1alpha1.Subscription{
+				in: &v1alpha1.Subscription{
 					ObjectMeta: metav1.ObjectMeta{
 						Name:      "sub",
 						Namespace: "ns",
@@ -1023,10 +996,10 @@ func TestInstallPlanReconcile(t *testing.T) {
 						},
 						LastUpdated: earlier,
 					},
-				}),
+				},
 			},
 			want: want{
-				out: newInstallPlanPendingState(&v1alpha1.Subscription{
+				out: &v1alpha1.Subscription{
 					ObjectMeta: metav1.ObjectMeta{
 						Name:      "sub",
 						Namespace: "ns",
@@ -1041,7 +1014,7 @@ func TestInstallPlanReconcile(t *testing.T) {
 						},
 						LastUpdated: earlier,
 					},
-				}),
+				},
 			},
 		},
 		{
@@ -1075,7 +1048,7 @@ func TestInstallPlanReconcile(t *testing.T) {
 				},
 			},
 			args: args{
-				in: newSubscriptionExistsState(&v1alpha1.Subscription{
+				in: &v1alpha1.Subscription{
 					ObjectMeta: metav1.ObjectMeta{
 						Name:      "sub",
 						Namespace: "ns",
@@ -1087,10 +1060,10 @@ func TestInstallPlanReconcile(t *testing.T) {
 						},
 						LastUpdated: earlier,
 					},
-				}),
+				},
 			},
 			want: want{
-				out: newInstallPlanPendingState(&v1alpha1.Subscription{
+				out: &v1alpha1.Subscription{
 					ObjectMeta: metav1.ObjectMeta{
 						Name:      "sub",
 						Namespace: "ns",
@@ -1105,7 +1078,7 @@ func TestInstallPlanReconcile(t *testing.T) {
 						},
 						LastUpdated: now,
 					},
-				}),
+				},
 			},
 		},
 		{
@@ -1139,7 +1112,7 @@ func TestInstallPlanReconcile(t *testing.T) {
 				},
 			},
 			args: args{
-				in: newSubscriptionExistsState(&v1alpha1.Subscription{
+				in: &v1alpha1.Subscription{
 					ObjectMeta: metav1.ObjectMeta{
 						Name:      "sub",
 						Namespace: "ns",
@@ -1151,10 +1124,10 @@ func TestInstallPlanReconcile(t *testing.T) {
 						},
 						LastUpdated: earlier,
 					},
-				}),
+				},
 			},
 			want: want{
-				out: newInstallPlanPendingState(&v1alpha1.Subscription{
+				out: &v1alpha1.Subscription{
 					ObjectMeta: metav1.ObjectMeta{
 						Name:      "sub",
 						Namespace: "ns",
@@ -1169,7 +1142,7 @@ func TestInstallPlanReconcile(t *testing.T) {
 						},
 						LastUpdated: now,
 					},
-				}),
+				},
 			},
 		},
 		{
@@ -1198,7 +1171,7 @@ func TestInstallPlanReconcile(t *testing.T) {
 				},
 			},
 			args: args{
-				in: newSubscriptionExistsState(&v1alpha1.Subscription{
+				in: &v1alpha1.Subscription{
 					ObjectMeta: metav1.ObjectMeta{
 						Name:      "sub",
 						Namespace: "ns",
@@ -1214,10 +1187,10 @@ func TestInstallPlanReconcile(t *testing.T) {
 						},
 						State: v1alpha1.SubscriptionStateUpgradePending,
 					},
-				}),
+				},
 			},
 			want: want{
-				out: newInstallPlanMissingState(&v1alpha1.Subscription{
+				out: &v1alpha1.Subscription{
 					ObjectMeta: metav1.ObjectMeta{
 						Name:      "sub",
 						Namespace: "ns",
@@ -1226,7 +1199,7 @@ func TestInstallPlanReconcile(t *testing.T) {
 						LastUpdated: now,
 						State:       v1alpha1.SubscriptionStateNone,
 					},
-				}),
+				},
 			},
 		},
 		{
@@ -1263,7 +1236,7 @@ func TestInstallPlanReconcile(t *testing.T) {
 				},
 			},
 			args: args{
-				in: newSubscriptionExistsState(&v1alpha1.Subscription{
+				in: &v1alpha1.Subscription{
 					ObjectMeta: metav1.ObjectMeta{
 						Name:      "sub",
 						Namespace: "ns",
@@ -1278,10 +1251,10 @@ func TestInstallPlanReconcile(t *testing.T) {
 						},
 						LastUpdated: earlier,
 					},
-				}),
+				},
 			},
 			want: want{
-				out: newInstallPlanPendingState(&v1alpha1.Subscription{
+				out: &v1alpha1.Subscription{
 					ObjectMeta: metav1.ObjectMeta{
 						Name:      "sub",
 						Namespace: "ns",
@@ -1296,7 +1269,7 @@ func TestInstallPlanReconcile(t *testing.T) {
 						},
 						LastUpdated: earlier,
 					},
-				}),
+				},
 			},
 		},
 		{
@@ -1333,7 +1306,7 @@ func TestInstallPlanReconcile(t *testing.T) {
 				},
 			},
 			args: args{
-				in: newSubscriptionExistsState(&v1alpha1.Subscription{
+				in: &v1alpha1.Subscription{
 					ObjectMeta: metav1.ObjectMeta{
 						Name:      "sub",
 						Namespace: "ns",
@@ -1348,10 +1321,10 @@ func TestInstallPlanReconcile(t *testing.T) {
 						},
 						LastUpdated: earlier,
 					},
-				}),
+				},
 			},
 			want: want{
-				out: newInstallPlanFailedState(&v1alpha1.Subscription{
+				out: &v1alpha1.Subscription{
 					ObjectMeta: metav1.ObjectMeta{
 						Name:      "sub",
 						Namespace: "ns",
@@ -1366,7 +1339,7 @@ func TestInstallPlanReconcile(t *testing.T) {
 						},
 						LastUpdated: now,
 					},
-				}),
+				},
 			},
 		},
 		{
@@ -1410,7 +1383,7 @@ func TestInstallPlanReconcile(t *testing.T) {
 				},
 			},
 			args: args{
-				in: newSubscriptionExistsState(&v1alpha1.Subscription{
+				in: &v1alpha1.Subscription{
 					ObjectMeta: metav1.ObjectMeta{
 						Name:      "sub",
 						Namespace: "ns",
@@ -1425,10 +1398,10 @@ func TestInstallPlanReconcile(t *testing.T) {
 						},
 						LastUpdated: earlier,
 					},
-				}),
+				},
 			},
 			want: want{
-				out: newInstallPlanFailedState(&v1alpha1.Subscription{
+				out: &v1alpha1.Subscription{
 					ObjectMeta: metav1.ObjectMeta{
 						Name:      "sub",
 						Namespace: "ns",
@@ -1443,7 +1416,7 @@ func TestInstallPlanReconcile(t *testing.T) {
 						},
 						LastUpdated: now,
 					},
-				}),
+				},
 			},
 		},
 		{
@@ -1487,7 +1460,7 @@ func TestInstallPlanReconcile(t *testing.T) {
 				},
 			},
 			args: args{
-				in: newSubscriptionExistsState(&v1alpha1.Subscription{
+				in: &v1alpha1.Subscription{
 					ObjectMeta: metav1.ObjectMeta{
 						Name:      "sub",
 						Namespace: "ns",
@@ -1502,10 +1475,10 @@ func TestInstallPlanReconcile(t *testing.T) {
 						},
 						LastUpdated: earlier,
 					},
-				}),
+				},
 			},
 			want: want{
-				out: newInstallPlanFailedState(&v1alpha1.Subscription{
+				out: &v1alpha1.Subscription{
 					ObjectMeta: metav1.ObjectMeta{
 						Name:      "sub",
 						Namespace: "ns",
@@ -1520,7 +1493,7 @@ func TestInstallPlanReconcile(t *testing.T) {
 						},
 						LastUpdated: earlier,
 					},
-				}),
+				},
 			},
 		},
 		{
@@ -1557,7 +1530,7 @@ func TestInstallPlanReconcile(t *testing.T) {
 				},
 			},
 			args: args{
-				in: newSubscriptionExistsState(&v1alpha1.Subscription{
+				in: &v1alpha1.Subscription{
 					ObjectMeta: metav1.ObjectMeta{
 						Name:      "sub",
 						Namespace: "ns",
@@ -1572,10 +1545,10 @@ func TestInstallPlanReconcile(t *testing.T) {
 						},
 						LastUpdated: earlier,
 					},
-				}),
+				},
 			},
 			want: want{
-				out: newInstallPlanInstalledState(&v1alpha1.Subscription{
+				out: &v1alpha1.Subscription{
 					ObjectMeta: metav1.ObjectMeta{
 						Name:      "sub",
 						Namespace: "ns",
@@ -1585,9 +1558,9 @@ func TestInstallPlanReconcile(t *testing.T) {
 							Namespace: "ns",
 							Name:      "ip",
 						},
-						LastUpdated: now,
+						LastUpdated: earlier,
 					},
-				}),
+				},
 			},
 		},
 		{
@@ -1621,7 +1594,7 @@ func TestInstallPlanReconcile(t *testing.T) {
 				},
 			},
 			args: args{
-				in: newSubscriptionExistsState(&v1alpha1.Subscription{
+				in: &v1alpha1.Subscription{
 					ObjectMeta: metav1.ObjectMeta{
 						Name:      "sub",
 						Namespace: "ns",
@@ -1633,10 +1606,10 @@ func TestInstallPlanReconcile(t *testing.T) {
 						},
 						LastUpdated: earlier,
 					},
-				}),
+				},
 			},
 			want: want{
-				out: newInstallPlanInstalledState(&v1alpha1.Subscription{
+				out: &v1alpha1.Subscription{
 					ObjectMeta: metav1.ObjectMeta{
 						Name:      "sub",
 						Namespace: "ns",
@@ -1648,7 +1621,7 @@ func TestInstallPlanReconcile(t *testing.T) {
 						},
 						LastUpdated: earlier,
 					},
-				}),
+				},
 			},
 		},
 	}
@@ -1663,12 +1636,6 @@ func TestInstallPlanReconcile(t *testing.T) {
 			out, err := rec.Reconcile(ctx, tt.args.in)
 			require.Equal(t, tt.want.err, err)
 			require.Equal(t, tt.want.out, out)
-
-			// Ensure the client's view of the subscription matches the typestate's
-			sub := out.(SubscriptionState).Subscription()
-			clusterSub, err := rec.client.OperatorsV1alpha1().Subscriptions(sub.GetNamespace()).Get(context.TODO(), sub.GetName(), metav1.GetOptions{})
-			require.NoError(t, err)
-			require.Equal(t, sub, clusterSub)
 		})
 	}
 }
@@ -1709,7 +1676,6 @@ func newFakeCatalogHealthReconciler(ctx context.Context, t *testing.T, config *f
 
 	rec := &catalogHealthReconciler{
 		now:                       config.now,
-		client:                    fakeClient,
 		catalogLister:             lister.OperatorsV1alpha1().CatalogSourceLister(),
 		registryReconcilerFactory: config.registryReconcilerFactory,
 		globalCatalogNamespace:    config.globalCatalogNamespace,
@@ -1730,7 +1696,6 @@ func newFakeInstallPlanReconciler(ctx context.Context, t *testing.T, config *fak
 
 	rec := &installPlanReconciler{
 		now:               config.now,
-		client:            fakeClient,
 		installPlanLister: lister.OperatorsV1alpha1().InstallPlanLister(),
 	}
 
@@ -1738,81 +1703,6 @@ func newFakeInstallPlanReconciler(ctx context.Context, t *testing.T, config *fak
 	versionedFactory.WaitForCacheSync(ctx.Done())
 
 	return rec
-}
-
-// Helper functions to shortcut to a particular state.
-// They should not be used outside of testing.
-
-func newSubscriptionExistsState(sub *v1alpha1.Subscription) SubscriptionExistsState {
-	return &subscriptionExistsState{
-		SubscriptionState: NewSubscriptionState(sub),
-	}
-}
-
-func newCatalogHealthState(sub *v1alpha1.Subscription) CatalogHealthState {
-	return &catalogHealthState{
-		SubscriptionExistsState: newSubscriptionExistsState(sub),
-	}
-}
-
-func newCatalogHealthKnownState(sub *v1alpha1.Subscription) CatalogHealthKnownState {
-	return &catalogHealthKnownState{
-		CatalogHealthState: newCatalogHealthState(sub),
-	}
-}
-
-func newCatalogHealthyState(sub *v1alpha1.Subscription) CatalogHealthyState {
-	return &catalogHealthyState{
-		CatalogHealthKnownState: newCatalogHealthKnownState(sub),
-	}
-}
-
-func newCatalogUnhealthyState(sub *v1alpha1.Subscription) CatalogUnhealthyState {
-	return &catalogUnhealthyState{
-		CatalogHealthKnownState: newCatalogHealthKnownState(sub),
-	}
-}
-
-func newNoInstallPlanReferencedState(sub *v1alpha1.Subscription) NoInstallPlanReferencedState {
-	return &noInstallPlanReferencedState{
-		InstallPlanState: newInstallPlanState(newSubscriptionExistsState(sub)),
-	}
-}
-
-func newInstallPlanReferencedState(sub *v1alpha1.Subscription) InstallPlanReferencedState {
-	return &installPlanReferencedState{
-		InstallPlanState: newInstallPlanState(newSubscriptionExistsState(sub)),
-	}
-}
-
-func newInstallPlanKnownState(sub *v1alpha1.Subscription) InstallPlanKnownState {
-	return &installPlanKnownState{
-		InstallPlanReferencedState: newInstallPlanReferencedState(sub),
-	}
-}
-
-func newInstallPlanMissingState(sub *v1alpha1.Subscription) InstallPlanMissingState {
-	return &installPlanMissingState{
-		InstallPlanKnownState: newInstallPlanKnownState(sub),
-	}
-}
-
-func newInstallPlanPendingState(sub *v1alpha1.Subscription) InstallPlanPendingState {
-	return &installPlanPendingState{
-		InstallPlanKnownState: newInstallPlanKnownState(sub),
-	}
-}
-
-func newInstallPlanFailedState(sub *v1alpha1.Subscription) InstallPlanFailedState {
-	return &installPlanFailedState{
-		InstallPlanKnownState: newInstallPlanKnownState(sub),
-	}
-}
-
-func newInstallPlanInstalledState(sub *v1alpha1.Subscription) InstallPlanInstalledState {
-	return &installPlanInstalledState{
-		InstallPlanKnownState: newInstallPlanKnownState(sub),
-	}
 }
 
 // Helper functions for generating OLM resources.
@@ -1849,4 +1739,44 @@ func withInstallPlanStatus(plan *v1alpha1.InstallPlan, status *v1alpha1.InstallP
 	plan.Status = *status
 
 	return plan
+}
+
+func catalogHealth(namespace, name string, lastUpdated *metav1.Time, healthy bool) v1alpha1.SubscriptionCatalogHealth {
+	return v1alpha1.SubscriptionCatalogHealth{
+		CatalogSourceRef: &corev1.ObjectReference{
+			Kind:       v1alpha1.CatalogSourceKind,
+			Namespace:  namespace,
+			Name:       name,
+			UID:        types.UID(name),
+			APIVersion: v1alpha1.CatalogSourceCRDAPIVersion,
+		},
+		LastUpdated: lastUpdated,
+		Healthy:     healthy,
+	}
+}
+
+func subscriptionCondition(conditionType v1alpha1.SubscriptionConditionType, status corev1.ConditionStatus, reason, message string, time *metav1.Time) v1alpha1.SubscriptionCondition {
+	return v1alpha1.SubscriptionCondition{
+		Type:               conditionType,
+		Status:             status,
+		Reason:             reason,
+		Message:            message,
+		LastTransitionTime: time,
+	}
+}
+
+func catalogUnhealthyCondition(status corev1.ConditionStatus, reason, message string, time *metav1.Time) v1alpha1.SubscriptionCondition {
+	return subscriptionCondition(v1alpha1.SubscriptionCatalogSourcesUnhealthy, status, reason, message, time)
+}
+
+func planMissingCondition(status corev1.ConditionStatus, reason, message string, time *metav1.Time) v1alpha1.SubscriptionCondition {
+	return subscriptionCondition(v1alpha1.SubscriptionInstallPlanMissing, status, reason, message, time)
+}
+
+func planFailedCondition(status corev1.ConditionStatus, reason, message string, time *metav1.Time) v1alpha1.SubscriptionCondition {
+	return subscriptionCondition(v1alpha1.SubscriptionInstallPlanFailed, status, reason, message, time)
+}
+
+func planPendingCondition(status corev1.ConditionStatus, reason, message string, time *metav1.Time) v1alpha1.SubscriptionCondition {
+	return subscriptionCondition(v1alpha1.SubscriptionInstallPlanPending, status, reason, message, time)
 }
