+++ pkg/controller/operators/catalog/subscriptions_test.go
@@ -46,7 +46,7 @@ func TestSyncSubscriptions(t *testing.T) {
 		existingOLMObjs      []runtime.Object
 	}
 	type args struct {
-		obj interface{}
+		obj *v1alpha1.Subscription
 	}
 	tests := []struct {
 		name              string
@@ -56,13 +56,6 @@ func TestSyncSubscriptions(t *testing.T) {
 		wantInstallPlans  []v1alpha1.InstallPlan
 		wantSubscriptions []*v1alpha1.Subscription
 	}{
-		{
-			name: "BadObject",
-			args: args{
-				obj: &v1alpha1.ClusterServiceVersion{},
-			},
-			wantErr: fmt.Errorf("casting Subscription failed"),
-		},
 		{
 			name: "NoStatus/NoCurrentCSV/MissingCatalogSourceNamespace",
 			fields: fields{
@@ -1313,11 +1306,11 @@ func TestSyncSubscriptions(t *testing.T) {
 					Name: testNamespace,
 				},
 			}
-			if err := o.syncSubscriptions(tt.args.obj); err != nil {
-				require.Equal(t, tt.wantErr, err)
-			} else {
-				require.Equal(t, tt.wantErr, o.syncResolvingNamespace(namespace))
-			}
+			//if _, err := o.syncSubscriptions(ctx, tt.args.obj); err != nil {
+			//	require.Equal(t, tt.wantErr, err)
+			//} else {
+			require.Equal(t, tt.wantErr, o.syncResolvingNamespace(namespace))
+			//}
 
 			for _, s := range tt.wantSubscriptions {
 				fetched, err := o.client.OperatorsV1alpha1().Subscriptions(testNamespace).Get(context.TODO(), s.GetName(), metav1.GetOptions{})
