+++ pkg/controller/registry/resolver/step_resolver.go
@@ -2,7 +2,6 @@
 package resolver
 
 import (
-	"context"
 	"fmt"
 
 	"github.com/operator-framework/api/pkg/operators/v1alpha1"
@@ -16,7 +15,6 @@ import (
 	"github.com/sirupsen/logrus"
 	corev1 "k8s.io/api/core/v1"
 	"k8s.io/apimachinery/pkg/api/errors"
-	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 	"k8s.io/apimachinery/pkg/labels"
 )
 
@@ -91,7 +89,7 @@ func NewOperatorStepResolver(lister operatorlister.OperatorLister, client versio
 }
 
 func (r *OperatorStepResolver) ResolveSteps(namespace string) ([]*v1alpha1.Step, []v1alpha1.BundleLookup, []*v1alpha1.Subscription, error) {
-	subs, err := r.listSubscriptions(namespace)
+	subs, err := r.subLister.Subscriptions(namespace).List(labels.Everything())
 	if err != nil {
 		return nil, nil, nil, err
 	}
@@ -119,9 +117,9 @@ func (r *OperatorStepResolver) ResolveSteps(namespace string) ([]*v1alpha1.Step,
 
 	// if there's no error, we were able to satisfy all constraints in the subscription set, so we calculate what
 	// changes to persist to the cluster and write them out as `steps`
-	steps := []*v1alpha1.Step{}
-	updatedSubs := []*v1alpha1.Subscription{}
-	bundleLookups := []v1alpha1.BundleLookup{}
+	var steps []*v1alpha1.Step
+	var updatedSubs []*v1alpha1.Subscription
+	var bundleLookups []v1alpha1.BundleLookup
 	for _, op := range operators {
 		// Find any existing subscriptions that resolve to this operator.
 		existingSubscriptions := make(map[*v1alpha1.Subscription]bool)
@@ -219,8 +217,9 @@ func (r *OperatorStepResolver) ResolveSteps(namespace string) ([]*v1alpha1.Step,
 				continue
 			}
 			// update existing subscription status
-			sub.Status.CurrentCSV = op.Name
-			updatedSubs = append(updatedSubs, sub)
+			updatedSub := sub.DeepCopy()
+			updatedSub.Status.CurrentCSV = op.Name
+			updatedSubs = append(updatedSubs, updatedSub)
 		}
 	}
 
@@ -243,19 +242,19 @@ func (r *OperatorStepResolver) hasExistingCurrentCSV(sub *v1alpha1.Subscription)
 	return false, err // Can't answer this question right now.
 }
 
-func (r *OperatorStepResolver) listSubscriptions(namespace string) ([]*v1alpha1.Subscription, error) {
-	list, err := r.client.OperatorsV1alpha1().Subscriptions(namespace).List(context.TODO(), metav1.ListOptions{})
-	if err != nil {
-		return nil, err
-	}
-
-	var subs []*v1alpha1.Subscription
-	for i := range list.Items {
-		subs = append(subs, &list.Items[i])
-	}
-
-	return subs, nil
-}
+//func (r *OperatorStepResolver) listSubscriptions(namespace string) ([]*v1alpha1.Subscription, error) {
+//	list, err := r.client.OperatorsV1alpha1().Subscriptions(namespace).List(context.TODO(), metav1.ListOptions{})
+//	if err != nil {
+//		return nil, err
+//	}
+//
+//	var subs []*v1alpha1.Subscription
+//	for i := range list.Items {
+//		subs = append(subs, &list.Items[i])
+//	}
+//
+//	return subs, nil
+//}
 
 type mergedSourceProvider struct {
 	sps    []cache.SourceProvider
